requirements.txt
Pillow
rembg